pyproject.toml
[tool.poetry]
name = "pipsqueak3"
version = "3.2.2"
description = "Mechasqueak3"
authors = ["Joshua Salzedo <thHunkn0WNd@gmail.com>"]
license = "BSD-3-Clause"
packages = [
    { include = "src"}
]

[tool.poetry.dependencies]
python = ">=3.8, <=4.0"
aiohttp = ">=3.4.0"
pydle = ">=0.9.1"
pure-sasl = "*"
humanfriendly = "*"
psycopg2-binary = "*"
toml = "*"
loguru = "*"
pluggy = "*"
async_lru = "^1.0.2"
prometheus_client = "^0.7.1"
prometheus_async = "^19.2.0"
pyparsing = "^2.4.7"
graypy = "^2.1.0"
python-dateutil = "^2.8.1"
websockets = "^8.1"
cattrs = ">=1.0.0"
pendulum = "^2.1.2"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-cov = "*"
pytest-asyncio = "*"
pylint = "*"
pytest-httpserver = "*"
codecov = "*"
pycodestyle = "*"
hypothesis = "^5.8.0"
ipython = "^7.18.1"
black = "^20.8b1"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
